change ADC sample rate to 1kHz
this costs us about 9% of our CPU, but should make aliasing much less
of a problem.

+++ libraries/AP_ADC/AP_ADC_HIL.h
@@ -36,7 +36,7 @@ class AP_ADC_HIL : public AP_ADC
 
 	///
 	// Read 6 sensors at once
-	uint16_t Ch6(const uint8_t *channel_numbers, uint16_t *result);
+	uint32_t Ch6(const uint8_t *channel_numbers, uint16_t *result);
 
 	///
 	// Set the adc raw values given the current rotations rates,
